Added listener to platform to permanently update android app state.
